test/wpt-test-page.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>WPT File System Tests (via polyfill)</title>
  <script src="/wpt/resources/testharness.js"></script>
</head>
<body>
  <div id="log"></div>
  <script>
    // Configure the WPT test harness
    setup({
      explicit_timeout: true,
      explicit_done: true,
    });

    // Store results globally so Playwright can access them
    window.__wptResults = null;
    add_completion_callback(function(tests, harnessStatus) {
      window.__wptResults = tests.map(function(t) {
        return {
          name: t.name,
          status: t.status,
          message: t.message || null,
          stack: t.stack || null,
        };
      });
    });
  </script>
  <script type="module">
    // Import the polyfill
    import * as fs from '/src/es6.js';
    const { getOriginPrivateDirectory } = fs;

    // Expose polyfill classes as globals (needed by some WPT test scripts)
    Object.assign(window, fs);

    // Get the test scripts to run from the URL query string
    // e.g. ?scripts=FileSystemDirectoryHandle-getFileHandle.js,FileSystemDirectoryHandle-removeEntry.js
    const params = new URLSearchParams(location.search);
    const scriptNames = (params.get('scripts') || '').split(',').filter(Boolean);
    const scriptDir = params.get('dir') || 'fs';

    // ──────────────────────────────────────────────────────
    // Set up test helpers as globals (matching WPT test-helpers.js)
    // ──────────────────────────────────────────────────────

    const kCurrentDirectory = '.';
    const kParentDirectory = '..';
    const kPathSeparators = ['/', '\\'];

    async function getFileSize(handle) {
      const file = await handle.getFile();
      return file.size;
    }

    async function getFileContents(handle) {
      const file = await handle.getFile();
      return new Response(file).text();
    }

    async function getDirectoryEntryCount(handle) {
      let result = 0;
      for await (const entry of handle) {
        result++;
      }
      return result;
    }

    async function getSortedDirectoryEntries(handle) {
      let result = [];
      for await (const entry of handle.values()) {
        if (entry.kind === 'directory') {
          result.push(entry.name + '/');
        } else {
          result.push(entry.name);
        }
      }
      result.sort();
      return result;
    }

    async function createDirectory(name, parent) {
      return await parent.getDirectoryHandle(name, { create: true });
    }

    async function createEmptyFile(name, parent) {
      const handle = await parent.getFileHandle(name, { create: true });
      assert_equals(await getFileSize(handle), 0);
      return handle;
    }

    async function createFileWithContents(name, contents, parent) {
      const handle = await createEmptyFile(name, parent);
      const writer = await handle.createWritable();
      await writer.write(new Blob([contents]));
      await writer.close();
      return handle;
    }

    async function cleanup(test, value, cleanup_func) {
      test.add_cleanup(async () => {
        try { await cleanup_func(); } catch (e) { /* ignore */ }
      });
      return value;
    }

    async function cleanup_writable(test, value) {
      return cleanup(test, value, async () => {
        try { await value.close(); } catch (e) { /* ignore */ }
      });
    }

    function createFileHandles(dir, ...fileNames) {
      return Promise.all(
        fileNames.map(fileName => dir.getFileHandle(fileName, { create: true }))
      );
    }

    function createDirectoryHandles(dir, ...dirNames) {
      return Promise.all(
        dirNames.map(dirName => dir.getDirectoryHandle(dirName, { create: true }))
      );
    }

    // Cleanup sandboxed filesystem
    async function cleanupSandboxedFileSystem(root) {
      for await (const [name, entry] of root) {
        await root.removeEntry(name, { recursive: entry.kind === 'directory' });
      }
    }

    // ──────────────────────────────────────────────────────
    // Set up directory_test using the polyfill's memory adapter
    // ──────────────────────────────────────────────────────

    const root = await getOriginPrivateDirectory(import('/src/adapters/memory.js'));

    function directory_test(func, description) {
      promise_test(async t => {
        await cleanupSandboxedFileSystem(root);
        t.add_cleanup(async () => {
          await cleanupSandboxedFileSystem(root);
        });
        await func(t, root);
      }, description);
    }

    // Expose everything as globals for WPT test scripts
    Object.assign(window, {
      kCurrentDirectory,
      kParentDirectory,
      kPathSeparators,
      getFileSize,
      getFileContents,
      getDirectoryEntryCount,
      getSortedDirectoryEntries,
      createDirectory,
      createEmptyFile,
      createFileWithContents,
      cleanup,
      cleanup_writable,
      createFileHandles,
      createDirectoryHandles,
      directory_test,
      garbageCollect: () => {},
    });

    // ──────────────────────────────────────────────────────
    // Load WPT test scripts
    // ──────────────────────────────────────────────────────

    for (const scriptName of scriptNames) {
      try {
        await import(`/wpt/${scriptDir}/script-tests/${scriptName}`);
      } catch (e) {
        // Report import errors as test failures
        test(() => { throw e; }, `Import: ${scriptName}`);
      }
    }

    done();
  </script>
</body>
</html>
